+++ codes/llm/tokens.py
@@ -1,0 +1,1 @@
+erniebot_access_token = '************************************'
